Disable default stderr handling
This plugin implicitly claims stdout and stderr handling. SL 4.3 has a built-in callback for such linters. 

Since for this linter the actual lint errors to parse are on stderr, we either need to disable the standard handling, like proposed, or switch `error_stream = util.STREAM_STDERR`. 

Whatever we do, the plugin should in-the-long-run report configuration and other hard errors to the user using `self.notify_failure()`.